许可协议更新

+++ OberoniaAureaGene/MIsc/Dialog_CreateDiscriminatGene.cs
@@ -133,7 +133,7 @@ public class Dialog_CreateDiscriminatGene : GeneCreationDialogBase
                 {
                     continue;
                 }
-                float num2 = 34f + GeneSize.x + 4f * (float)(genepack.GeneSet.GenesListForReading.Count + 2);
+                float num2 = 34f + GeneSize.x + 4f * (genepack.GeneSet.GenesListForReading.Count + 2);
                 if (curX + num2 > rect.width - 16f)
                 {
                     curX = 4f;
@@ -189,7 +189,7 @@ public class Dialog_CreateDiscriminatGene : GeneCreationDialogBase
                 {
                     continue;
                 }
-                float num2 = 34f + GeneSize.x * (float)genepack.GeneSet.GenesListForReading.Count + 4f * (float)(genepack.GeneSet.GenesListForReading.Count + 2);
+                float num2 = 34f + GeneSize.x * genepack.GeneSet.GenesListForReading.Count + 4f * (genepack.GeneSet.GenesListForReading.Count + 2);
                 if (curX + num2 > rect.width - 16f)
                 {
                     curX = 4f;
